requirements.txt
fastapi
uvicorn[standard]
python-multipart
pytesseract
opencv-python
Pillow
pdf2image
langdetect
Jinja2